Agent/requirements.txt
google-genai
python-dotenv
fastapi
uvicorn[standard]
pydantic
groq
pymongo
